% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fss.R
\name{ens_dfss}
\alias{ens_dfss}
\title{Title}
\usage{
ens_dfss(
  x,
  threshold,
  radius,
  comparator = "ge",
  include_low = TRUE,
  include_high = TRUE,
  num_cores = 1
)
}
\arguments{
\item{num_cores}{}
}
\description{
Title
}
