% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hira_scores.R
\name{hira_scores}
\alias{hira_scores}
\title{Run "fuzzy" spatial verification for 1 case}
\usage{
hira_scores(score = NULL, execute = NULL, ...)
}
\arguments{
\item{obfield}{Observation grid.}

\item{fcfield}{Forecast field}

\item{thresholds}{A vector of thresholds}

\item{scales}{A vector of (odd!) window sizes}
}
\value{
A tibble with columns for threshold, window_size and various scores.
}
\description{
Run "fuzzy" spatial verification for 1 case
}
