% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hira_config.R
\name{read_hira_config}
\alias{read_hira_config}
\title{Read local harpSpatial configuration files}
\usage{
read_hira_config(config_file = NULL)
}
\arguments{
\item{config_file}{The name of the config file to be read.
If NULL, the current configuration is printed.}
}
\description{
Read local harpSpatial configuration files
}
