% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAL.R
\name{SAL}
\alias{SAL}
\title{SAL score}
\usage{
SAL(
  fcfield,
  obfield,
  thresh_scale = 15,
  min_rain = 0.1,
  same_threshold = FALSE,
  maxobj = 1000,
  add_objs = FALSE,
  ...
)
}
\arguments{
\item{fcfield}{The forecast field}

\item{obfield}{The observation field}

\item{thresh_scale}{Used for threshold of object boundaries: Rmax/threshscale}

\item{min_rain}{The minimum value to be considered. If no grid point is above, SAL is not computed.}

\item{same_threshold}{If TRUE, threshold is a single number for both fields}

\item{maxobj}{Maximum number of objects to identify}

\item{add_objs}{If TRUE, the object matrices are returned as attributes of the result.}

\item{...}{Ignored options.}

\item{threshold}{A 2-vector containing (precipitation) thresholds for fc and obs.}
}
\value{
a data.frame with S, A, L values.
}
\description{
SAL score
}
