% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_hira.R
\name{save_hira_verif}
\alias{save_hira_verif}
\title{Save spatial scores to SQLite}
\usage{
save_hira_verif(score_tables, sqlite_path, sqlite_file)
}
\arguments{
\item{sqlite_path}{The path for the sqlite file}

\item{sqlite_file}{The file to which the tables are written or added}

\item{scores}{A list of spatial score tables}
}
\description{
Save spatial scores to SQLite
}
