% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_spatial.R
\name{save_spatial_verif}
\alias{save_spatial_verif}
\title{Save spatial scores to SQLite}
\usage{
save_spatial_verif(scores, sqlite_path, sqlite_file)
}
\arguments{
\item{scores}{A list of spatial score tables}

\item{sqlite_path}{The path for the sqlite file}

\item{sqlite_file}{The file to which the tables are written or added}
}
\description{
Save spatial scores to SQLite
}
