% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_scores.R
\name{score_sp_aggregated}
\alias{score_sp_aggregated}
\title{Run spatial verification for 1 case}
\usage{
score_sp_aggregated(obfield, fcfield, ...)
}
\arguments{
\item{obfield}{Observation grid.}

\item{fcfield}{Forecast field}

\item{...}{ignored
return A 1-row tibble of scores}
}
\description{
Run spatial verification for 1 case
}
