% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hira_scores.R
\name{scores_hira}
\alias{scores_hira}
\title{Run "HiRA" spatial verification for 1 case}
\usage{
scores_hira(obsvect, indices, fcfield, thresholds, scales, strategies, ...)

scores_hira(obsvect, indices, fcfield, thresholds, scales, strategies, ...)
}
\arguments{
\item{obsvect}{One of HiRA Scores}

\item{fcvect}{If True then fire the score function}
}
\value{
A list of bias, mse and mae.
not exported

A list of bias, mse and mae.
not exported
}
\description{
Run "HiRA" spatial verification for 1 case

Run "HiRA" spatial verification for 1 case
}
