% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hira_scores.R
\name{scores_hira_basic}
\alias{scores_hira_basic}
\title{Run "HiRA" basic scores for 1 case}
\usage{
scores_hira_basic(obsvect, indices, fcfield, scales, ...)
}
\arguments{
\item{obsvect}{One of HiRA Scores}

\item{fcvect}{If True then fire the score function}
}
\value{
A list of bias, mse and mae.
not exported
}
\description{
Run "HiRA" basic scores for 1 case
}
