% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slx.R
\name{slx}
\alias{slx}
\title{SLX Score Calculation wrapper for verify_spatial}
\usage{
slx(obfield, fcfield, scales, ...)
}
\arguments{
\item{obfield}{A geofield object representing the observation field.}

\item{fcfield}{A geofield object representing the forecast field.}

\item{scales}{A vector of scales (in grid points) to compute the score for.}

\item{...}{Other arguments, not used.}
}
\value{
A tibble with columns for scale, S, L, and X.
}
\description{
This function calculates the SLX score for a given forecast and observation field.
It acts as a wrapper around the original SLX_components function to be compatible
with the verify_spatial workflow.
}
