% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_tables.R
\name{spatial_score_table}
\alias{spatial_score_table}
\title{return table description for spatial verification score tables}
\usage{
spatial_score_table(tab)
}
\description{
return table description for spatial verification score tables
}
