% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_scores.R
\name{verify_basic}
\alias{verify_basic}
\title{Run spatial verification for 1 case}
\usage{
verify_basic(obfield, fcfield)
}
\arguments{
\item{obfield}{Observation grid.}

\item{fcfield}{Forecast field
return A 1-row tibble of scores}
}
\description{
Run spatial verification for 1 case
}
