% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_scores.R
\name{verify_fuzzy}
\alias{verify_fuzzy}
\title{Run "fuzzy" spatial verification for 1 case}
\usage{
verify_fuzzy(obfield, fcfield, thresholds, window_sizes)
}
\arguments{
\item{obfield}{Observation grid.}

\item{fcfield}{Forecast field}

\item{thresholds}{A vector of thresholds}

\item{window_sizes}{A vector of (odd!) window sizes}
}
\value{
A tibble with columns for threshold, window_size and various scores.
}
\description{
Run "fuzzy" spatial verification for 1 case
}
