% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_hira.R
\name{verify_hira}
\alias{verify_hira}
\title{Run spatial verification on a (for now) deterministic forecast}
\usage{
verify_hira(
  dttm,
  parameter,
  lead_time = harpSpatial_hira_conf$lead_time,
  lt_unit = harpSpatial_hira_conf$lt_unit,
  stations = harpCore::station_list,
  padding_i = 5,
  padding_j = 5,
  scores = NULL,
  interp_method = "nearest",
  obs_path = ".",
  obsfile_template = harpSpatial_hira_conf$obsfile_template,
  fcst_model = harpSpatial_hira_conf$fcst_model,
  fc_file_path = harpSpatial_hira_conf$fc_file_path,
  fc_file_template = harpSpatial_hira_conf$fc_file_template,
  fc_file_format = harpSpatial_hira_conf$fc_file_format,
  fc_file_opts = harpSpatial_hira_conf$fc_file_opts,
  fc_accumulation = harpSpatial_hira_conf$fc_accumulation,
  window_sizes = harpSpatial_hira_conf$window_sizes,
  thresholds = harpSpatial_hira_conf$thresholds,
  sqlite_path = harpSpatial_hira_conf$sqlite_path,
  sqlite_file = harpSpatial_hira_conf$sqlite_file,
  return_data = FALSE
)
}
\arguments{
\item{dttm}{A vector of date time strings to read. Can be in YYYYMMDD,
YYYYMMDDhh, YYYYMMDDhhmm, or YYYYMMDDhhmmss format. Can be numeric or
character. \code{\link[harpCore]{seq_dttm}} can be used to generate a
vector of equally spaced date-time strings.}

\item{parameter}{The parameters to read as a character vector.}

\item{lead_time}{The lead times to read as a numeric vector.
Should be in the units that are also used in fc_file_template.}

\item{lt_unit}{The unit used for lead_time. Can be "h" (hours), "m" (minutes), "s" (seconds)}

\item{stations}{The IDs of the stations to read from the files. By default
this is full list from harpCore::station_list, the stations outside the domain interior will be excluded.}

\item{padding_i}{Number of grid points to define the domain interior in the x direction.}

\item{padding_j}{Number of grid points to define the domain interior in the y direction.}

\item{scores}{HiRA HiRA and basic scores
me:    Multi Event
pragm: Pragramtic 
crss:  Conditional Square Root RPS
pph:   Theat Detection
bias:  Bias 
mse:   Mean Squared Error  
mae:   Mean Absolute Error}

\item{interp_method}{interpolation method from model grid}

\item{obs_path}{Path to the observation files}

\item{obsfile_template}{Template of observation files.}

\item{fcst_model}{The name of the (deterministic or EPS) model.}

\item{fc_file_path}{The top level path for the forecast files to read.}

\item{fc_file_template}{The file type to generate the template for. Can be
"harmoneps_grib", "harmeoneps_grib_fp", "harmoneps_grib_sfx", "meps_met",
"harmonie_grib", "harmonie_grib_fp", "harmone_grib_sfx", "vfld", "vobs", or
"fctable". If anything else is passed, it is returned unmodified. In this
case substitutions can be used. Available substitutions are \{YYYY\} for
year, \{MM\} for 2 digit month with leading zero, \{M\} for month with no
leading zero, and similarly \{DD\} or \{D\} for day, \{HH\} or \{H\} for
hour, \{mm\} or \{m\} for minute. Also \{LDTx\} for lead time and \{MBRx\}
for ensemble member where x is the length of the string including leading
zeros - can be omitted or 2, 3 or 4. Note that the full path to the file
will always be file_path/template.}

\item{fc_file_format}{The format of the files to read. Can be e.g. "fa" or "grib".}

\item{fc_file_opts}{A list with format-specific options for the reader function.}

\item{fc_accumulation}{The accumulation type of the forecast. This is only used for
accumulated parameters (e.g. precipitation). NULL signifies that the field is accumulated
from the start of the model run. Otherwise this should be a string containing a numerical value
and a time unit, e.g. "15m" or "1h".}

\item{window_sizes}{Scales used for fuzzy methods like FSS. A vector of box sizes.
All values must be odd integers (so the central point is really in the center of a box).}

\item{thresholds}{Thresholds used for FSS, ...}

\item{sqlite_path}{If specified, SQLite files are generated and written to
this directory.}

\item{sqlite_file}{Name of SQLite file.}

\item{return_data}{If TRUE, the result is returned as a list of tables.}

\item{...}{Not used at thispoint (more info to be added).}
}
\value{
A list containting tibbles for all scores.
}
\description{
Run spatial verification on a (for now) deterministic forecast
}
