% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hira_scores.R
\name{hira_scores}
\alias{hira_scores}
\title{Run "HiRA" spatial verification for 1 case}
\usage{
hira_scores(score = NULL, execute = NULL, ...)
}
\arguments{
\item{score}{One of HiRA Scores}

\item{execute}{If True then fire the score function}
}
\value{
A tibble with columns for obsvect, fcstvect fcfield, threshold, scales ...
not exported
}
\description{
Run "HiRA" spatial verification for 1 case
}
