% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAL.R
\name{SAL}
\alias{SAL}
\title{SAL score}
\usage{
SAL(fcfield, obfield, threshScale = 15, threshold = pmax(0.1,
  c(max(fcfield, na.rm = TRUE), max(obfield, na.rm = TRUE)))/threshScale,
  sameThreshold = FALSE, maxobj = 1000, min.rain = 1,
  add.objs = FALSE)
}
\arguments{
\item{fcfield}{The forecast field}

\item{obfield}{The observation field}

\item{threshold}{A 2-vector containing (precipitation) thresholds for fc and obs.}

\item{sameThreshold}{If TRUE, threshold is a single number for both fields}

\item{maxobj}{Maximum number of objects to identify}

\item{min.rain}{The minimum value to be considered}

\item{add.objs}{If TRUE, the object matrices are returned as attributes of the result.}

\item{threshscale}{Used for threshold}
}
\value{
a vector with S, A, L values.
}
\description{
SAL score
}
