% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_scores.R
\name{spatial_scores}
\alias{spatial_scores}
\title{Calculate spatial scores}
\usage{
spatial_scores(score = NULL, obfield = NULL, fcfield = NULL, ...)
}
\arguments{
\item{score}{The score to calculate. If NULL, the function will return a list of available scores.}

\item{obfield}{A matrix containing the observation field. If NULL, the function will return
the table structure for the specified score.}

\item{fcfield}{A matrix containing the forecast field. Must have the same dimension as obfield.}

\item{...}{Other options that may depend on the score (like scale, threshold, ...)}
}
\description{
Calculate spatial scores
}
