% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_ens_spatial.R, R/verify_spatial.R
\name{verify_spatial}
\alias{verify_spatial}
\title{Run spatial verification on a (for now) deterministic forecast}
\usage{
verify_spatial(
  start_date,
  end_date,
  parameter,
  model = harpSpatial_conf$model,
  lead_time = harpSpatial_conf$lead_time,
  lt_unit = harpSpatial_conf$lt_unit,
  by = harpSpatial_conf$by,
  scores = NULL,
  members = harpSpatial_conf$members,
  fc_file_path = harpSpatial_conf$fc_file_path,
  fc_file_template = harpSpatial_conf$fc_file_template,
  fc_file_format = harpSpatial_conf$fc_file_format,
  fc_file_options = harpSpatial_conf$fc_file_options,
  fc_domain = harpSpatial_conf$fc_domain,
  fc_interp_method = harpSpatial_conf$fc_interp_method,
  fc_accumulation = harpSpatial_conf$fc_accumulation,
  ob_file_path = harpSpatial_conf$ob_file_path,
  ob_file_template = harpSpatial_conf$ob_file_template,
  ob_file_format = harpSpatial_conf$ob_file_format,
  ob_file_options = harpSpatial_conf$ob_file_options,
  ob_domain = harpSpatial_conf$ob_domain,
  ob_interp_method = harpSpatial_conf$ob_interp_method,
  ob_accumulation = harpSpatial_conf$ob_accumulation,
  verif_domain = harpSpatial_conf$verif_domain,
  use_mask = harpSpatial_conf$use_mask,
  window_sizes = harpSpatial_conf$window_sizes,
  thresholds = harpSpatial_conf$thresholds,
  sqlite_path = harpSpatial_conf$sqlite_path,
  sqlite_file = harpSpatial_conf$sqlite_file,
  return_data = FALSE
)

verify_spatial(
  start_date,
  end_date,
  parameter,
  model = harpSpatial_conf$model,
  lead_time = harpSpatial_conf$lead_time,
  lt_unit = harpSpatial_conf$lt_unit,
  by = harpSpatial_conf$by,
  scores = NULL,
  members = harpSpatial_conf$members,
  fc_file_path = harpSpatial_conf$fc_file_path,
  fc_file_template = harpSpatial_conf$fc_file_template,
  fc_file_format = harpSpatial_conf$fc_file_format,
  fc_file_options = harpSpatial_conf$fc_file_options,
  fc_domain = harpSpatial_conf$fc_domain,
  fc_interp_method = harpSpatial_conf$fc_interp_method,
  fc_accumulation = harpSpatial_conf$fc_accumulation,
  ob_file_path = harpSpatial_conf$ob_file_path,
  ob_file_template = harpSpatial_conf$ob_file_template,
  ob_file_format = harpSpatial_conf$ob_file_format,
  ob_file_options = harpSpatial_conf$ob_file_options,
  ob_domain = harpSpatial_conf$ob_domain,
  ob_interp_method = harpSpatial_conf$ob_interp_method,
  ob_accumulation = harpSpatial_conf$ob_accumulation,
  verif_domain = harpSpatial_conf$verif_domain,
  use_mask = harpSpatial_conf$use_mask,
  window_sizes = harpSpatial_conf$window_sizes,
  thresholds = harpSpatial_conf$thresholds,
  sqlite_path = harpSpatial_conf$sqlite_path,
  sqlite_file = harpSpatial_conf$sqlite_file,
  return_data = FALSE
)
}
\arguments{
\item{start_date}{Date of the first forecast to read.}

\item{end_date}{Date of the last forecast to read.}

\item{parameter}{The parameters to read as a character vector.}

\item{model}{The name of the (deterministic or EPS) model.}

\item{lead_time}{The lead times to read as a numeric vector.
Should be in the units that are also used in fc_file_template.}

\item{lt_unit}{The unit used for lead_time. Can be "h" (hours), "m" (minutes), "s" (seconds)}

\item{by}{The time between forecasts. Should be a string of a number followed
by a letter, where the letter gives the units - may be "d" for days, "h" for
hours or "m" for minutes.}

\item{members}{The (numbers of the) ensemble members to read. While Netcdf and grib2 files 
can contain multiple members, for other formats we assume they are in separate files
(see also fc_file_template)}

\item{fc_file_path}{The top level path for the forecast files to read.}

\item{fc_file_template}{The file type to generate the template for. Can be
"harmoneps_grib", "harmeoneps_grib_fp", "harmoneps_grib_sfx", "meps_met",
"harmonie_grib", "harmonie_grib_fp", "harmone_grib_sfx", "vfld", "vobs", or
"fctable". If anything else is passed, it is returned unmodified. In this
case substitutions can be used. Available substitutions are \{YYYY\} for
year, \{MM\} for 2 digit month with leading zero, \{M\} for month with no
leading zero, and similarly \{DD\} or \{D\} for day, \{HH\} or \{H\} for
hour, \{mm\} or \{m\} for minute. Also \{LDTx\} for lead time and \{MBRx\}
for ensemble member where x is the length of the string including leading
zeros - can be omitted or 2, 3 or 4. Note that the full path to the file
will always be file_path/template.}

\item{fc_file_format}{The format of the files to read. Can be e.g. "fa" or "grib".}

\item{fc_file_options}{A list with format-specific options for the reader function.}

\item{fc_domain}{The forecast domain. If provided, the fc reading can be made faster
by not extracting domain information (format option meta).}

\item{fc_interp_method}{Interpolation method to be used when transforming a forecast
field to the verification grid.}

\item{fc_accumulation}{The accumulation type of the forecast. This is only used for
accumulated parameters (e.g. precipitation). NULL signifies that the field is accumulated
from the start of the model run. Otherwise this should be a string containing a numerical value
and a time unit, e.g. "15m" or "1h".}

\item{ob_file_path}{The top level path for the forecast files to read.}

\item{ob_file_template}{The file type to generate the template for. Can be
"harmoneps_grib", "harmeoneps_grib_fp", "harmoneps_grib_sfx", "meps_met",
"harmonie_grib", "harmonie_grib_fp", "harmone_grib_sfx", "vfld", "vobs", or
"fctable". If anything else is passed, it is returned unmodified. In this
case substitutions can be used. Available substitutions are {YYYY} for
year, \{MM\} for 2 digit month with leading zero, \{M\} for month with no
leading zero, and similarly \{DD\} or \{D\} for day, \{HH\} or \{H\} for
hour, \{mm\} or \{m\} for minute. Also \{LDTx\} for lead time and \{MBRx\}
for ensemble member where x is the length of the string including leading
zeros - can be omitted or 2, 3 or 4. Note that the full path to the file
will always be file_path/template.}

\item{ob_file_format}{The format of the files to read. Can be e.g. "hdf5" or "grib".}

\item{ob_file_options}{A list with format-specific options for the reader function.}

\item{ob_domain}{The observation domain. If provided, the obs reading can be made faster
by not extracting domain information (format option meta).}

\item{ob_interp_method}{Interpolation method to be used when transforming a forecast
field to the verification grid.}

\item{ob_accumulation}{The accumulation type of the observation (or reference). This is only used for
accumulated parameters (e.g. precipitation). NULL signifies that the field is accumulated
from the start of the model run. That is probably rare for observations. 
Otherwise this should be a string containing a numerical value
and a time unit, e.g. "15m" or "1h". "0" means an instantaneous value.}

\item{verif_domain}{A \code{geodomain} that defines the common verification grid.}

\item{window_sizes}{Scales used for fuzzy methods like FSS. A vector of box sizes.
All values must be odd integers (so the central point is really in the center of a box).}

\item{thresholds}{Thresholds used for FSS, ...}

\item{sqlite_path}{If specified, SQLite files are generated and written to
this directory.}

\item{sqlite_file}{Name of SQLite file.}

\item{return_data}{If TRUE, the result is returned as a list of tables.}

\item{fc_options}{A list with format-specific options for the reader function.}

\item{ob_options}{A list with format-specific options for the reader function.}

\item{...}{Not used at thispoint (more info to be added).}
}
\value{
A list containting tibbles for all scores.

A list containting tibbles for all scores.
}
\description{
Run spatial verification on a (for now) deterministic forecast

Run spatial verification on a (for now) deterministic forecast
}
