% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_spatial.R
\name{verify_spatial}
\alias{verify_spatial}
\title{Run spatial verification on a (for now) deterministic forecast}
\usage{
verify_spatial(start_date, end_date, det_model, parameter,
  lead_time = seq(0, 36, 3), lt_unit = "h", by = "6h",
  fc_file_path = "", fc_file_template = "", fc_file_format = "fatar",
  fc_options = list(), fc_interp_method = "closest",
  fc_accumulation = NULL, ob_file_path = "", ob_file_template = "",
  ob_file_format = "hdf5", ob_options = list(),
  ob_interp_method = "closest", ob_accumulation = "15m",
  verif_domain = NULL, use_mask = FALSE, box_sizes = c(1, 3, 5, 11,
  21), thresholds = c(0.1, 1, 5, 10), sqlite_path = NULL,
  return_data = TRUE, ...)
}
\arguments{
\item{start_date}{Date of the first forecast to read.}

\item{end_date}{Date of the last forecast to read.}

\item{det_model}{The name of the deterministic model. Maybe expressed as a
vector if more than one model is wanted.}

\item{parameter}{The parameters to read as a character vector. For reading
from vfld files, set to NULL to read all parameters.}

\item{lead_time}{The lead times to read as a numeric vector.}

\item{by}{The time between forecasts. Should be a string of a number followed
by a letter, where the letter gives the units - may be d for days, h for
hours or m for minutes.}

\item{use_mask}{If TRUE, a land/sea mask is used when interpolating. It must be
available in the forecast files or in a clim_file.}

\item{box_sizes}{Scales used for fuzzy methods like FSS. A vector of odd box sizes.}

\item{thresholds}{Thresholds used for FSS, ...}

\item{sqlite_path}{If specified, SQLite files are generated and written to
this directory.}

\item{...}{Arguments dependent on \code{file_format}, passed to to read_XXX_interpolate.
(More info to be added).}

\item{file_path}{The top level path for the forecast files to read.}

\item{file_format}{The format of the files to read. Can be "vfld", "grib" or
"netcdf_met"}

\item{file_template}{The file type to generate the template for. Can be
"harmoneps_grib", "harmeoneps_grib_fp", "harmoneps_grib_sfx", "meps_met",
"harmonie_grib", "harmonie_grib_fp", "harmone_grib_sfx", "vfld", "vobs", or
"fctable". If anything else is passed, it is returned unmodified. In this
case substitutions can be used. Available substitutions are {YYYY} for
year, \{MM\} for 2 digit month with leading zero, \{M\} for month with no
leading zero, and similarly \{DD\} or \{D\} for day, \{HH\} or \{H\} for
hour, \{mm\} or \{m\} for minute. Also \{LDTx\} for lead time and \{MBRx\}
for ensemble member where x is the length of the string including leading
zeros - can be omitted or 2, 3 or 4. Note that the full path to the file
will always be file_path/template.}

\item{clim_file}{A file containing constant data for the domain:
topology, land/sea mask.}

\item{clim_format}{The file format of the clim_file may be different
than that of the forecast files.}

\item{correct_T2m}{Whether to correct the 2m temperature forecast from the
model elevation to the observation elevation.}

\item{interpolation_method}{The method used for interpolating from forecast grid
to station points. Default is "closest" (mearest neighbour).
Alternatives include "bilin".}
}
\value{
A list containting two data frames: \code{spatial_scores} and
  \code{spatial_threshold_scores}.
}
\description{
Run spatial verification on a (for now) deterministic forecast
}
